Remove unused files related to visualisation
